fix files style-gide, pagination
