+++ src/js/catalog_page/catalog-refs.js
@@ -1,0 +1,8 @@
+export const refs = {
+    form: document.querySelector('#search-form'),
+    searchInput: document.querySelector('[name="searchQuery"]'),
+    gallery: document.querySelector('.gallery'),
+    clearBtn: document.querySelector('.clear-btn'),
+    yearSelectEl: document.getElementById('year-select'),
+    pageContainer: document.querySelector('.pagination ul'),
+};
