+++ src/partials/catalog_page/section_catalog.html
@@ -31,11 +31,11 @@
     <ul id="anchor" class="gallery">
       <!-- Карточки изображений -->
     </ul>
-    <div class="pagination">
+    <!-- <div class="pagination">
       <ul>
-        <!--pages or li are comes from javascript -->
+        pages or li are comes from javascript
       </ul>
-    </div>
-    <!-- <div id="tui-pagination-container" class="tui-pagination"></div> -->
+    </div> -->
+    <div id="tui-pagination-container" class="tui-pagination"></div>
   </div>
 </section>
